Move code for struct type into a sub-module
